SVN r34903 |2014-10-16 22:16:18 +0000

+++ SequenceAnalysis/src/org/labkey/sequenceanalysis/run/analysis/AASnpByCodonAggregator.java
@@ -321,7 +321,7 @@ public class AASnpByCodonAggregator extends NtSnpByPosAggregator
                 depth += getCoverageAggregator().getDepthAtPosition(_ntRefName, nt.first, 0); //always use depth at last non-indel position
             }
 
-            return (depth / _ntPositions.size());
+            return (depth / (double)_ntPositions.size());
         }
 
         public int getRefNtId()
